+++ lib/openmaize/confirm.ex
@@ -1,9 +1,9 @@
 defmodule Openmaize.Confirm do
   @moduledoc """
-  Module to handle email confirmation.
+  Module to help with email confirmation.
 
   See the documentation for Openmaize.Signup.add_confirm_token for details
-  about producing the confirmation token.
+  about creating the confirmation token.
   """
 
   import Comeonin.Tools
@@ -13,9 +13,8 @@ defmodule Openmaize.Confirm do
   @doc """
   Verify the token sent by email.
 
-  If the token is valid, `{:ok, user, email}` will be returned.
-
-  If the token is not valid, `{:error, message}` will be returned.
+  If the token is valid, `{:ok, user, email}` will be returned, and
+  if the token is not valid, `{:error, message}` will be returned.
 
   ## Options
 
@@ -24,10 +23,28 @@ defmodule Openmaize.Confirm do
   * query_function - a custom function to query the database
     * if you are using Ecto, you will probably not need this
 
-  This is a function that you need to provide. See the examples below
-  to see how to set this option when calling `user_email`.
-
   ## Examples
+
+  The example below shows a function for the `confirm` route, which
+  needs to be set in the `web/router.ex` file, in a controller for
+  a Phoenix app.
+
+  The `Mailer.receipt_confirm` function is a function that you need to
+  write to send an email stating that confirmation was successful.
+
+      def confirm(conn, params) do
+        case Openmaize.Confirm.user_email(conn) do
+          {:ok, user, email} ->
+            Mailer.receipt_confirm(email)
+            conn
+            |> put_flash(:info, "You have successfully confirmed your account.")
+            |> redirect(to: login_path(conn, :login))
+          {:error, message} ->
+            conn
+            |> put_flash(:error, "Something went wrong with the confirmation of your account.")
+            |> redirect(to: page_path(conn, :index))
+        end
+      end
 
   """
   def user_email(%Plug.Conn{params: %{"email" => email, "key" => key}}, opts \\ []) do